Added option '--mask <file>' (short form '-m <file>').
The new option allows to mask symbols using a list of regular expressions.

It behaves like '--whitelist <file>', except that that the file
contains a list of regular expressions (masks), not exact symbol names.
It helps to filter out C++ false-positives that typically contain '::'.
For example, 'std::[^:]*::' is an attempt to filter out false-positives
on STL class members.